# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/salons": {
      "post": {
        "operationId": "SalonsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSalonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      },
      "get": {
        "operationId": "SalonsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      }
    },
    "/salons/my-salon": {
      "get": {
        "operationId": "SalonsController_getMyCurrentSalon",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      }
    },
    "/salons/{id}": {
      "get": {
        "operationId": "SalonsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      },
      "patch": {
        "operationId": "SalonsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSalonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      },
      "delete": {
        "operationId": "SalonsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      }
    },
    "/salons/{id}/disable-users": {
      "patch": {
        "operationId": "SalonsController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      }
    },
    "/salons/{id}/enable-users": {
      "patch": {
        "operationId": "SalonsController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Salons"
        ]
      }
    },
...
  },
  "info": {
    "title": "Barberbook API",
    "description": "Documentación de la API de Barberbook",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Ingrese su token JWT en el siguiente formato: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateSalonDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSalonDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}